refactoring files and updated paths.
